fix -Wtype-limits for test-all.cpp (#29)

+++ callsign.h
@@ -32,7 +32,7 @@ public:
 	void CodeIn(const uint8_t *code);
 	const std::string GetCS(unsigned len = 0) const;
 	void CodeOut(uint8_t *out) const;
-	size_t Hash() const { return coded; }
+	uint64_t Hash() const { return coded; }
 	bool operator==(const CCallsign &rhs) const;
 	bool operator!=(const CCallsign &rhs) const;
 	char GetModule(void) const;
